inline nssb into nssdb_browser
